+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/PowerBankScreen.java
@@ -13,7 +13,6 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 
-//TODO correct values
 public class PowerBankScreen extends AbstractContainerScreen<PowerBankMenu> {
 
     public static final ResourceLocation TEXTURE = ResourceLocation.fromNamespaceAndPath(Stellaris.MODID, "textures/gui/power_bank.png");