+++ common/src/main/java/com/st0x0ef/stellaris/common/blocks/entities/machines/FuelRefineryBlockEntity.java
@@ -152,7 +152,6 @@ public class FuelRefineryBlockEntity extends BaseEnergyContainerBlockEntity impl
     public SingleFluidStorage getIngredientTank() {
         return inputTank;
     }
-
     public SingleFluidStorage getOutputFuelTank() {
         return outputFuelTank;
     }