+++ neoforge/src/main/java/com/st0x0ef/stellaris/neoforge/client/StellarisNeoforgeClient.java
@@ -134,6 +134,7 @@ public class StellarisNeoforgeClient {
         event.register(MenuTypesRegistry.TABLET_MENU.get(), TabletMainScreen::new);
         event.register(MenuTypesRegistry.DIESEL_GENERATOR_MENU.get(), DieselGeneratorScreen::new);
         event.register(MenuTypesRegistry.POWER_BANK_MENU.get(), PowerBankScreen::new);
+        event.register(MenuTypesRegistry.FLUID_TANK_MENU.get(), FluidTankScreen::new);
     }
 
     @SubscribeEvent